Use double data type for the SPM atlas affine
UniSeg errored for me because the input affine I gave was torch32. think it is best to use double data type for the SPM atlas affine matrix because it is what is returned by default when reading with io.map().affine, plus it is better for numerical precision without much memory overhead.

+++ nitorch/tools/uniseg/_api.py
@@ -237,7 +237,7 @@ def get_spm_prior(**backend):
     f = io.map(fname).movedim(-1, 0)[:-1]  # drop background
     aff = f.affine
     dat = f.fdata(**backend)
-    aff = aff.to(**utils.backend(dat))
+    aff = aff.to(**utils.backend(dat)).type(torch.float64)
     return dat, aff
